requirements.txt
jupyter-client==7.4.7
jupyter-core==5.0.0
matplotlib-inline==0.1.6
numpy==1.23.5
pandas==1.5.2
plotly==5.11.0
pydeck==0.8.0
streamlit==1.18.1
transformers==4.26.0
sentence-transformers==2.2.2
postgrest-py==0.10.3
SQLAlchemy==1.4.46
scikit-learn==1.2.1
scipy==1.10.0
keybert==0.7.0
streamlit-option-menu==0.3.2
jwt==1.3.1
torch==1.13.1
extra-streamlit-components==0.1.56
watchdog==2.2.1
joblib==1.2.0
Pillow==9.4.0
psycopg2-binary==2.9.5
pyarrow==11.0.0
torchmetrics==0.11.1
torchvision==0.14.1
xgboost==1.7.3